3D breadboard test original file uploads
These files contain the prefabricated game assets for the breadboard and the wire, as well as all the two use. The code uses a different technique to manipulate data between game objects and does not resemble the 2D test files in many ways. Future game objects will be uploaded here. Prefabricated game assets should be able to be simply imported to a project in unity and be available to be placed in scene immediately.